+++ src/apscheduler/datastores/sqlalchemy.py
@@ -317,7 +317,6 @@ class SQLAlchemyDataStore(BaseExternalDataStore):
             Column("start_deadline", timestamp_type),
             Column("result_expiration_time", interval_type),
             Column("created_at", timestamp_type, nullable=False),
-            Column("started_at", timestamp_type),
             Column("acquired_by", Unicode(500)),
             Column("acquired_until", timestamp_type),
         )
@@ -326,7 +325,8 @@ class SQLAlchemyDataStore(BaseExternalDataStore):
             metadata,
             Column("job_id", Uuid, primary_key=True),
             Column("outcome", Enum(JobOutcome), nullable=False),
-            Column("finished_at", timestamp_type, index=True),
+            Column("started_at", timestamp_type, index=True),
+            Column("finished_at", timestamp_type, nullable=False, index=True),
             Column("expires_at", timestamp_type, nullable=False, index=True),
             Column("exception", LargeBinary),
             Column("return_value", LargeBinary),